Multi branch role maturity generation
This adds the feature to do multiple branches in the role
maturity generation. If the user gives a --branch flag, for
example like --branch stable/ocata, the generation will try
to checkout openstack-ansible and all the openstack-ansible.*
governance roles into their stable/ocata branch to find a
role maturity file there.

In this example, if there is no stable/ocata branch,
the role is ommited. If the stable/ocata branch has no
openstack-ansible metadata, the role maturity will be marked
as unknown, as usual.