Add EOL support
When looking for a reference (like stable/newton) for its
latest sha, a reference that is not found will be returned as is
(new branches, safe behavior of no modification).

If a branch is considered EOL, its reference won't exist as a
branch anymore, but a tag with the name '-eol' (without the 'stable/'
prefix) exists. Here we are looking for that exception, and if a
tag for EOL exist, return the tag instead of its SHA.

It's more explicit for end-users.